Making a service for fetching all the prices
The Albion wordpress theme uses the shopify buy sdk, which is deprecated
and can't fetch prices based on the country. We need this for the Global-E
dynamic pricing thing.

I could have written this in PHP but I CBA working out how to use
Composer on the wordpress server so am making this microservice thing

this is just the PoC